refactor(ui): reorganize `shared/utils.ts`
- its main purpose is for namespace manipulation, so rename to `shared/namespaces.ts`
  - and rename `Utils` to `nsUtils` for explicitness as well
    - this singleton may not be necessary in the future, but keeping it as-is right now for simpler diff
  - generally speaking, "dumping grounds" are not good to have and organizing them when they get large is best practice

- several util functions were used in exactly one place, so just move them to that place
  - `statusIconClasses` was only used in `phase-icon.tsx`
  - `isWorkflowRunning`, `isWorkflowSuspended` were only used in `workflow-operations-map.tsx`

- some were only used together with exactly one component, so move to that component
  - `getValueFromParameter` was only used with `ParametersInput`

- others were only used for one section of the UI, e.g. Events or Workflows
  - so move them to a `utils.ts` file there; they are not shared between sections
  - also `queryParams` was only used in `services/`

- also simplify some of these functions:
  - with optional chaining
  - by inverting unnecesarily nested conditionals
  - by removing unnecessary singleton structures

- these changes also make these functions easier to unit test and give them proper encapsulation
  - private functions are no longer exported
