+++ ui/src/app/plugins/plugin-list.tsx
@@ -7,15 +7,15 @@ import {uiUrl} from '../shared/base';
 import {useCollectEvent} from '../shared/use-collect-event';
 import {ZeroState} from '../shared/components/zero-state';
 import {historyUrl} from '../shared/history';
-import {Utils} from '../shared/utils';
+import {nsUtils} from '../shared/namespaces';
 
 export function PluginList({match, history}: RouteComponentProps<any>) {
     // state for URL and query parameters
-    const [namespace] = useState(Utils.getNamespace(match.params.namespace) || '');
+    const [namespace] = useState(nsUtils.getNamespace(match.params.namespace) || '');
     useEffect(
         () =>
             history.push(
-                historyUrl('plugins' + (Utils.managedNamespace ? '' : '/{namespace}'), {
+                historyUrl('plugins' + (nsUtils.managedNamespace ? '' : '/{namespace}'), {
                     namespace
                 })
             ),
