+++ ui/src/app/shared/components/namespace-filter.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import {Utils} from '../utils';
+import {nsUtils} from '../namespaces';
 import {InputFilter} from './input-filter';
 
 export const NamespaceFilter = (props: {value: string; onChange: (namespace: string) => void}) =>
-    Utils.managedNamespace ? <>{Utils.managedNamespace}</> : <InputFilter value={props.value} name='ns' onChange={ns => props.onChange(ns)} />;
+    nsUtils.managedNamespace ? <>{nsUtils.managedNamespace}</> : <InputFilter value={props.value} name='ns' onChange={ns => props.onChange(ns)} />;
