+++ ui/src/app/shared/history.test.ts
@@ -2,12 +2,12 @@
  * @jest-environment jsdom
  */
 import {historyUrl} from './history';
-import {Utils} from './utils';
+import {nsUtils} from './namespaces';
 
 describe('history URL', () => {
     test('namespace', () => {
         expect(historyUrl('foo/{namespace}', {namespace: 'my-ns'})).toBe('/foo/my-ns?');
-        expect(Utils.currentNamespace).toBe('my-ns');
+        expect(nsUtils.currentNamespace).toBe('my-ns');
     });
 
     test('path parameter', () => {
