+++ ui/src/app/shared/services/workflows-service.ts
@@ -6,7 +6,7 @@ import {ResubmitOpts, RetryOpts} from '../../../models';
 import {SubmitOpts} from '../../../models/submit-opts';
 import {uiUrl} from '../base';
 import {Pagination} from '../pagination';
-import {Utils} from '../utils';
+import {queryParams} from './utils';
 import requests from './requests';
 import {WorkflowDeleteResponse} from './responses';
 
@@ -53,7 +53,7 @@ export const WorkflowsService = {
             'items.spec.suspend'
         ]
     ) {
-        const params = Utils.queryParams({phases, labels, pagination});
+        const params = queryParams({phases, labels, pagination});
         params.push(`fields=${fields.join(',')}`);
         return requests.get(`api/v1/workflows/${namespace}?${params.join('&')}`).then(res => res.body as WorkflowList);
     },
@@ -73,7 +73,7 @@ export const WorkflowsService = {
         labels?: Array<string>;
         resourceVersion?: string;
     }): Observable<models.kubernetes.WatchEvent<Workflow>> {
-        const url = `api/v1/workflow-events/${query.namespace || ''}?${Utils.queryParams(query).join('&')}`;
+        const url = `api/v1/workflow-events/${query.namespace || ''}?${queryParams(query).join('&')}`;
         return requests.loadEventSource(url).pipe(map(data => data && (JSON.parse(data).result as models.kubernetes.WatchEvent<Workflow>)));
     },
 
@@ -90,7 +90,7 @@ export const WorkflowsService = {
         labels?: Array<string>;
         resourceVersion?: string;
     }): Observable<models.kubernetes.WatchEvent<Workflow>> {
-        const params = Utils.queryParams(query);
+        const params = queryParams(query);
         const fields = [
             'result.object.metadata.name',
             'result.object.metadata.namespace',
