+++ ui/src/app/shared/workflow-operations-map.tsx
@@ -1,7 +1,6 @@
 import type {NodePhase, Workflow} from '../../models';
 import {services} from './services';
 import {WorkflowDeleteResponse} from './services/responses';
-import {Utils} from './utils';
 
 export type OperationDisabled = {
     [action in WorkflowOperationName]: boolean;
@@ -41,25 +40,25 @@ export const WorkflowOperationsMap: WorkflowOperations = {
     SUSPEND: {
         title: 'SUSPEND',
         iconClassName: 'fa fa-pause',
-        disabled: (wf: Workflow) => !Utils.isWorkflowRunning(wf) || Utils.isWorkflowSuspended(wf),
+        disabled: (wf: Workflow) => !isWorkflowRunning(wf) || isWorkflowSuspended(wf),
         action: (wf: Workflow) => services.workflows.suspend(wf.metadata.name, wf.metadata.namespace)
     },
     RESUME: {
         title: 'RESUME',
         iconClassName: 'fa fa-play',
-        disabled: (wf: Workflow) => !Utils.isWorkflowSuspended(wf),
+        disabled: (wf: Workflow) => !isWorkflowSuspended(wf),
         action: (wf: Workflow) => services.workflows.resume(wf.metadata.name, wf.metadata.namespace, null)
     },
     STOP: {
         title: 'STOP',
         iconClassName: 'fa fa-stop-circle',
-        disabled: (wf: Workflow) => !Utils.isWorkflowRunning(wf),
+        disabled: (wf: Workflow) => !isWorkflowRunning(wf),
         action: (wf: Workflow) => services.workflows.stop(wf.metadata.name, wf.metadata.namespace)
     },
     TERMINATE: {
         title: 'TERMINATE',
         iconClassName: 'fa fa-times-circle',
-        disabled: (wf: Workflow) => !Utils.isWorkflowRunning(wf),
+        disabled: (wf: Workflow) => !isWorkflowRunning(wf),
         action: (wf: Workflow) => services.workflows.terminate(wf.metadata.name, wf.metadata.namespace)
     },
     DELETE: {
@@ -69,3 +68,27 @@ export const WorkflowOperationsMap: WorkflowOperations = {
         action: (wf: Workflow) => services.workflows.delete(wf.metadata.name, wf.metadata.namespace)
     }
 };
+
+function isWorkflowSuspended(wf: Workflow): boolean {
+    if (!wf?.spec) {
+        return false;
+    }
+    if (wf.spec.suspend) {
+        return true;
+    }
+    if (wf.status?.nodes) {
+        for (const node of Object.values(wf.status.nodes)) {
+            if (node.type === 'Suspend' && node.phase === 'Running') {
+                return true;
+            }
+        }
+    }
+    return false;
+}
+
+function isWorkflowRunning(wf: Workflow): boolean {
+    if (!wf?.spec) {
+        return false;
+    }
+    return wf.status.phase === 'Running';
+}
