+++ ui/src/app/workflows/components/workflows-list/workflows-list.tsx
@@ -20,7 +20,7 @@ import {ListWatch, sortByYouth} from '../../../shared/list-watch';
 import {Pagination, parseLimit} from '../../../shared/pagination';
 import {ScopedLocalStorage} from '../../../shared/scoped-local-storage';
 import {services} from '../../../shared/services';
-import {Utils} from '../../../shared/utils';
+import {nsUtils} from '../../../shared/namespaces';
 import * as Actions from '../../../shared/workflow-operations-map';
 import {WorkflowCreator} from '../workflow-creator';
 import {WorkflowFilters} from '../workflow-filters/workflow-filters';
@@ -53,7 +53,7 @@ export function WorkflowsList({match, location, history}: RouteComponentProps<an
     const queryParams = new URLSearchParams(location.search);
     const {navigation} = useContext(Context);
 
-    const [namespace, setNamespace] = useState(Utils.getNamespace(match.params.namespace) || '');
+    const [namespace, setNamespace] = useState(nsUtils.getNamespace(match.params.namespace) || '');
     const [pagination, setPagination] = useState<Pagination>(() => {
         const savedPaginationLimit = storage.getItem('options', {}).paginationLimit || undefined;
         return {
@@ -133,7 +133,7 @@ export function WorkflowsList({match, location, history}: RouteComponentProps<an
         if (pagination.limit) {
             params.append('limit', pagination.limit.toString());
         }
-        history.push(historyUrl('workflows' + (Utils.managedNamespace ? '' : '/{namespace}'), {namespace, extraSearchParams: params}));
+        history.push(historyUrl('workflows' + (nsUtils.managedNamespace ? '' : '/{namespace}'), {namespace, extraSearchParams: params}));
     }, [namespace, phases.toString(), labels.toString(), pagination.limit, pagination.offset]); // referential equality, so use values, not refs
 
     useEffect(() => {
@@ -317,7 +317,7 @@ export function WorkflowsList({match, location, history}: RouteComponentProps<an
             <SlidingPanel isShown={!!getSidePanel()} onClose={() => navigation.goto('.', {sidePanel: null})}>
                 {getSidePanel() === 'submit-new-workflow' && (
                     <WorkflowCreator
-                        namespace={Utils.getNamespaceWithDefault(namespace)}
+                        namespace={nsUtils.getNamespaceWithDefault(namespace)}
                         onCreate={wf => navigation.goto(uiUrl(`workflows/${wf.metadata.namespace}/${wf.metadata.name}`))}
                     />
                 )}
