+++ ui/src/app/workflows/utils.ts
@@ -1,0 +1,3 @@
+export function shortNodeName(node: {name: string; displayName: string}): string {
+    return node.displayName || node.name;
+}
